requirements.txt
streamlit
uuid
pydantic
python-dotenv
google-generativeai
google-genai
